タスク一覧画面に表示する日時のフォーマットをjsファイルに移動
- タスク日時を編集する際に使用するフォーマット関数を明示的な名前に修正
