+++ src/pages/Home.jsx
@@ -5,6 +5,7 @@ import axios from 'axios';
 import { Header } from '../components/Header';
 import { url } from '../const';
 import './home.scss';
+import { formatForDisplay } from '../dateFormat';
 
 export const Home = () => {
   const [isDoneDisplay, setIsDoneDisplay] = useState('todo'); // todo->未完了 done->完了
@@ -115,18 +116,6 @@ export const Home = () => {
 
 // 表示するタスク
 const Tasks = (props) => {
-  const formatDate = (isoDateString) => {
-    const date = new Date(isoDateString);
-    date.setHours(date.getHours() - 9);
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-    const hours = String(date.getHours()).padStart(2, '0');
-    const minutes = String(date.getMinutes()).padStart(2, '0');
-
-    return `${year}/${month}/${day} ${hours}:${minutes}`;
-  };
-
   const { tasks, selectListId, isDoneDisplay } = props;
   if (tasks === null) return <></>;
 
@@ -138,7 +127,7 @@ const Tasks = (props) => {
             return task.done === true;
           })
           .map((task, key) => {
-            const limit = formatDate(task.limit);
+            const limit = formatForDisplay(task.limit);
             return (
               <li key={key} className="task-item">
                 <Link to={`/lists/${selectListId}/tasks/${task.id}`} className="task-item-link">
@@ -161,7 +150,7 @@ const Tasks = (props) => {
           return task.done === false;
         })
         .map((task, key) => {
-          const limit = formatDate(task.limit);
+          const limit = formatForDisplay(task.limit);
 
           return (
             <li key={key} className="task-item">
